# frontend/Sources/Networking/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vca-api",
    "description": "value-compass FastAPI service. Reads from Postgres only; never calls Polygon directly except via the /portfolio/holdings background task for brand-new tickers.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/portfolio/status": {
      "get": {
        "summary": "Portfolio Status",
        "description": "Return the freshest ``last_modified`` / ``next_modified``.\n\nThe cache response is lightweight and Cloudflare-cacheable.",
        "operationId": "portfolio_status_portfolio_status_get",
        "parameters": [
          {
            "name": "X-App-Attest",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Attest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioStatusResponse"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "description": "Cloudflare-cacheable max-age in seconds.",
                "schema": {
                  "type": "string"
                }
              },
              "Last-Modified": {
                "description": "UTC RFC 7231 timestamp for the response.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Min-App-Version": {
                "description": "Minimum supported iOS app version.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid App Attest credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "description": "Cloudflare-cacheable max-age in seconds.",
                "schema": {
                  "type": "string"
                }
              },
              "Last-Modified": {
                "description": "UTC RFC 7231 timestamp for the response.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Min-App-Version": {
                "description": "Minimum supported iOS app version.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "description": "Cloudflare-cacheable max-age in seconds.",
                "schema": {
                  "type": "string"
                }
              },
              "Last-Modified": {
                "description": "UTC RFC 7231 timestamp for the response.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Min-App-Version": {
                "description": "Minimum supported iOS app version.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "appAttestMissing",
          "conflictDetected",
          "lossyMappingRejected",
          "portfolioNotFound",
          "schemaUnsupported",
          "stockDataMissing",
          "stockDataPending",
          "stockDataStale",
          "syncUnavailable",
          "unsupportedMovingAverageWindow"
        ],
        "title": "ErrorCode",
        "description": "Stable machine-readable error codes for clients."
      },
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "description": "Structured HTTP error body decoded by iOS clients."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioStatusResponse": {
        "properties": {
          "last_modified": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Modified"
          },
          "next_modified": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Modified"
          }
        },
        "type": "object",
        "required": [
          "last_modified",
          "next_modified"
        ],
        "title": "PortfolioStatusResponse",
        "description": "Portfolio freshness response body."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}